feat: create WeekPageTable component and used week switcher on teamSlug/week.page.tsx
