+++ app/pages/[teamSlug]/time/week.page.tsx
@@ -1,121 +1,43 @@
-import { addDays, differenceInDays, eachDayOfInterval, format, formatISO, isToday, parseISO } from 'date-fns'
-import { useRouter } from 'next/router'
-import { DayWeekSwitch } from '../../../frontend/components/dayWeekSwitchButton'
-import { FormattedDuration } from '../../../frontend/components/duration/formattedDuration'
-import { HourInput } from '../../../frontend/components/hourInput'
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeadCell,
-  TableHeadRow,
-  TableRow,
-} from '../../../frontend/components/table/table'
-import { ProjectFragment, TaskFragment, useWorkHoursQuery } from '../../../frontend/generated/graphql'
+import { addDays, endOfWeek, format, parse, startOfWeek } from "date-fns"
+import Link from "next/link"
+import { useRouter } from "next/router"
+import { WeekPageTable } from "../../../frontend/components/WeekPageTable"
 
-const NUMBER_OF_DAYS = 7
-
-interface WorkHoursTableRow {
-  task: TaskFragment
-  project: ProjectFragment
-  durations: number[]
-}
-const WeekPage = () => {
-  const fromDate = new Date(2022, 7, 8)
-  const toDate = addDays(fromDate, NUMBER_OF_DAYS - 1)
-  const interval = { start: fromDate, end: toDate }
+const WeekPage = (): JSX.Element => {
+  
   const router = useRouter()
-  const [{ data }] = useWorkHoursQuery({
-    variables: {
-      teamSlug: router.query.teamSlug?.toString() ?? '',
-      from: formatISO(fromDate, { representation: 'date' }),
-      to: formatISO(toDate, { representation: 'date' }),
-    },
-  })
-  const tableData: WorkHoursTableRow[] = []
-  for (const workHour of data?.workHours ?? []) {
-    const workHourDate = parseISO(workHour.date)
-    const weekDay = differenceInDays(workHourDate, fromDate)
-    const rowIndex = tableData.findIndex((row) => {
-      return row.task.id === workHour.task.id
-    })
-    if (rowIndex === -1) {
-      tableData.push({
-        project: workHour.project,
-        task: workHour.task,
-        durations: eachDayOfInterval(interval).map(() => 0),
-      })
-      tableData[tableData.length - 1].durations[weekDay] = workHour.duration
-    } else {
-      tableData[rowIndex].durations[weekDay] += workHour.duration
-    }
-  }
+const teamSlug = router.query.teamSlug
+  const urlDate = router.query.date?.toString()
 
-  const classNameMarkDay = 'bg-slate-300 dark:bg-gray-900'
+  const currentDate = urlDate ? parse(urlDate, 'yyyy-MM-dd', new Date()) : new Date()
 
-  return (
-    <div>
-      <Table>
-        <TableHead>
-          <TableHeadRow>
-            <TableHeadCell />
-            {eachDayOfInterval(interval).map((day) => (
-              <TableHeadCell className={isToday(day) ? classNameMarkDay : ''} key={day.toString()}>
-                {format(day, 'EEE')}
-                <br />
-                {format(day, 'dd. MMM')}
-              </TableHeadCell>
-            ))}
-            <TableHeadCell />
-          </TableHeadRow>
-        </TableHead>
-        <TableBody>
-          {tableData.map((row) => (
-            <TableRow key={row.task.id}>
-              <TableCell>
-                {row.project.title}
-                <br />
-                {row.task.title}
-              </TableCell>
-              {eachDayOfInterval(interval).map((day, dayIndex) => (
-                <TableCell className={isToday(day) ? classNameMarkDay : ''} key={day.toString()}>
-                  <HourInput readOnly onChange={console.log} workHours={row.durations[dayIndex] / 60} />
-                </TableCell>
-              ))}
-              <TableCell>
-                <FormattedDuration
-                  title=""
-                  minutes={row.durations.reduce((previousValue, currentValue) => previousValue + currentValue)}
-                />
-              </TableCell>
-            </TableRow>
-          ))}
-          <TableRow>
-            <TableCell />
-            {eachDayOfInterval(interval).map((day, dayIndex) => (
-              <TableCell className={isToday(day) ? classNameMarkDay : ''} key={day.toString()}>
-                <FormattedDuration
-                  title=""
-                  minutes={tableData
-                    .map((row) => row.durations[dayIndex])
-                    .reduce((previousValue, currentValue) => previousValue + currentValue, 0)}
-                />
-              </TableCell>
-            ))}
-            <TableCell>
-              <FormattedDuration
-                title=""
-                minutes={tableData
-                  .flatMap((row) => row.durations)
-                  .reduce((previousValue, currentValue) => previousValue + currentValue, 0)}
-              />
-            </TableCell>
-          </TableRow>
-        </TableBody>
-      </Table>
-      <DayWeekSwitch selectedButton="week" />
-    </div>
-  )
-}
+  const startOfTheWeek = startOfWeek(currentDate, { weekStartsOn: 1 }) //startOfWeek in camel case doesn´t work, because of the following function: startOfWeek()
+
+  const endOfTheWeek = endOfWeek(currentDate, { weekStartsOn: 1 })
+
+  const startDate = format(startOfTheWeek, 'dd.MM')
+  const endDate = format(endOfTheWeek, 'dd.MM.yyyy')
+  
+
+
+  const nextWeek = addDays(currentDate, 7)
+  const previousWeek = addDays(currentDate, -7)
+  const previousWeekString = format(previousWeek, 'yyyy-MM-dd')
+  const nextWeekString = format(nextWeek, 'yyyy-MM-dd')
+return (
+  <>
+  
+  <WeekPageTable startDate={startOfTheWeek} />
+  <span className="text-center text-xs">
+    <Link href={`/${teamSlug}/time/week?date=${previousWeekString}`}>
+      <a className={`k w-10 rounded-l-lg bg-gray-400 px-2 py-1`}>Last</a>
+    </Link>
+    <Link href={`/${teamSlug}/time/week?date=${nextWeekString}`}>
+      <a className={` w-10 rounded-r-lg bg-gray-400 px-2 py-1`}>Next</a>
+    </Link>
+  </span>
+    <div>{startDate}-{endDate}</div>
+</>
+
+)}
 export default WeekPage
